fix tests, move constants
