+++ frontend/src/components/MeetingRoom/Toolbar/Toolbar.spec.tsx
@@ -3,6 +3,7 @@ import { render, screen } from '@testing-library/react';
 import { useLocation } from 'react-router-dom';
 import useSpeakingDetector from '../../../hooks/useSpeakingDetector';
 import Toolbar, { ToolbarProps } from './Toolbar';
+import isReportIssueEnabled from '../../../utils/isReportIssueEnabled';
 import useToolbarCount from '../../../hooks/useToolbarCount';
 
 const mockedRoomName = { roomName: 'test-room-name' };
@@ -15,9 +16,11 @@ vi.mock('react-router-dom', () => ({
 
 vi.mock('../../../hooks/useSpeakingDetector');
 vi.mock('../../../hooks/useToolbarCount');
+vi.mock('../../../utils/isReportIssueEnabled');
 
 const mockUseSpeakingDetector = useSpeakingDetector as Mock<[], boolean>;
 const mockUseToolbarCount = useToolbarCount as Mock<[], number>;
+const mockIsReportIssueEnabled = isReportIssueEnabled as Mock<[], boolean>;
 
 describe('Toolbar', () => {
   beforeEach(() => {
@@ -26,6 +29,7 @@ describe('Toolbar', () => {
     });
     mockUseSpeakingDetector.mockReturnValue(false);
     mockUseToolbarCount.mockReturnValue(9);
+    mockIsReportIssueEnabled.mockReturnValue(false);
   });
   afterAll(() => {
     vi.restoreAllMocks();
@@ -42,18 +46,12 @@ describe('Toolbar', () => {
   };
 
   it('does not render the Report Issue button if it is configured to be disabled', () => {
-    vi.stubEnv('VITE_ENABLE_REPORT_ISSUE', 'false');
-    render(<Toolbar {...defaultProps} />);
-    expect(screen.queryByTestId('report-issue-button')).not.toBeInTheDocument();
-  });
-
-  it('does not render the Report Issue button if the configuration is not defined', () => {
     render(<Toolbar {...defaultProps} />);
     expect(screen.queryByTestId('report-issue-button')).not.toBeInTheDocument();
   });
 
   it('renders the Report Issue button if it is configured to be enabled', () => {
-    vi.stubEnv('VITE_ENABLE_REPORT_ISSUE', 'true');
+    mockIsReportIssueEnabled.mockReturnValue(true);
     render(<Toolbar {...defaultProps} />);
     expect(screen.getByTestId('report-issue-button')).toBeInTheDocument();
   });
