+++ frontend/src/utils/isReportIssueEnabled/index.tsx
@@ -1,0 +1,3 @@
+import isReportIssueEnabled from './isReportIssueEnabled';
+
+export default isReportIssueEnabled;
